+++ src/lib/internal/actions/interact-outside/types.ts
@@ -10,8 +10,17 @@ export type InteractOutsideInterceptEventType =
 	| 'click';
 
 export type InteractOutsideInterceptHandler<E extends InteractOutsideInterceptEventType> = (
-	ev: HTMLElementEventMap[E]
+	ev: HTMLElementEventMap[E],
+	computedEventData?: ComputedEventData
 ) => void;
+
+/**
+ * Some event data is lost when the event is delayed, debounced or throttled
+ * as described in the dom-spec - https://dom.spec.whatwg.org/#event-path
+ * For this reason we can pre-save any needed data and store
+ * it here before being passed on the related events
+ * */
+export type ComputedEventData = { shadowTarget?: EventTarget };
 
 export type InteractOutsideConfig = {
 	/**
@@ -19,7 +28,7 @@ export type InteractOutsideConfig = {
 	 * which is either a `pointerup`, `mouseup`, or `touchend`
 	 * event, depending on the user's input device.
 	 */
-	onInteractOutside?: (e: InteractOutsideEvent) => void;
+	onInteractOutside?: (e: InteractOutsideEvent, computedEventData?: ComputedEventData) => void;
 
 	/**
 	 * Callback fired when an outside interaction event starts,