+++ src/lib/internal/helpers/is.ts
@@ -10,6 +10,10 @@ export function isDocument(element: unknown): element is Document {
 	return element instanceof Document;
 }
 
+export function isShadowRoot(element: unknown): element is ShadowRoot {
+	return element instanceof ShadowRoot;
+}
+
 export function isElement(element: unknown): element is Element {
 	return element instanceof Element;
 }
@@ -78,4 +82,18 @@ export function isReadable(value: unknown): value is Readable<unknown> {
 
 export function isWritable(value: unknown): value is Writable<unknown> {
 	return isReadable(value) && 'set' in value;
+}
+
+export function isAttachedToDocument(element: unknown): boolean {
+	if (!isHTMLElement(element)) return false;
+
+	const rootNode = element.getRootNode();
+
+	if (rootNode === document) return true;
+
+	if (isShadowRoot(rootNode)) {
+		return rootNode.ownerDocument === document;
+	}
+
+	return false;
 }